TASK: Extract package `Neos.Fusion.FluidAdapter`
The Fusion Prototype `Neos.Fusion:Template` is deprecated together with the `fusion:render` view-helper.
Both are moved into the new package `Neos.Fusion.FluidAdapter`.
