+++ Neos.Fusion.FluidAdapter/Configuration/Settings.yaml
@@ -1,0 +1,5 @@
+Neos:
+  Neos:
+    fusion:
+      autoInclude:
+        Neos.Fusion.FluidAdapter: true
