+++ Neos.Fusion.FluidAdapter/Readme.rst
@@ -1,0 +1,25 @@
+------
+Fusion.FluidAdapter
+------
+
+.. note:: This repository is a **read-only subsplit** of a package that is part of the
+          Neos project (learn more on `www.neos.io <https://www.neos.io/>`_).
+
+This is an adaptor for use of Fluid with Fusion.
+
+Fluid is a templating framework originally developed for Flow and Neos, now developed
+independently at https://github.com/typo3/fluid.
+
+Fusion is a hierarchical, prototype based processing language used in
+Neos. Documentation is `available as part of the Neos documentation
+<http://neos.readthedocs.org/en/stable/CreatingASite/Fusion/index.html>`_.
+
+If you want to use Neos, please have a look at the `Neos documentation
+<http://neos.readthedocs.org/en/stable/>`_
+
+Contribute
+----------
+
+If you want to contribute to Neos, please have a look at
+https://github.com/neos/neos-development-collection - it is the repository
+used for development and all pull requests should go into it.
