+++ Neos.Fusion.FluidAdapter/Resources/Private/Fusion/Root.fusion
@@ -1,0 +1,1 @@
+include: **/*.fusion
