+++ Neos.Fusion.FluidAdapter/Tests/Functional/FusionObjects/Fixtures/Fusion/Root.fusion
@@ -1,0 +1,1 @@
+include: ./**/*.fusion
