+++ Neos.Fusion.FluidAdapter/Tests/Functional/FusionObjects/Fixtures/Layouts/TestLayout.html
@@ -1,0 +1,1 @@
+layout start -- <f:render section="main" /> -- layout end
