+++ Neos.Fusion.FluidAdapter/Tests/Functional/FusionObjects/Fixtures/Partials/TestPartial.html
@@ -1,0 +1,1 @@
+--partial contents
