+++ Neos.Fusion.FluidAdapter/Tests/Functional/FusionObjects/Fixtures/Templates/JsonEncodedValueTemplate.html
