--- Neos.Fusion/Resources/Private/Fusion/Prototypes/Template.fusion
@@ -1,7 +1,0 @@
-prototype(Neos.Fusion:Template) {
-  @class = 'Neos\\Fusion\\FusionObjects\\TemplateImplementation'
-  templatePath = null
-  partialRootPath = null
-  layoutRootPath = null
-  sectionName = null
-}
