+++ Neos.Neos/Classes/ViewHelpers/Rendering/InPreviewModeViewHelper.php
@@ -16,7 +16,7 @@ namespace Neos\Neos\ViewHelpers\Rendering;
 
 use Neos\ContentRepository\Core\Projection\ContentGraph\Node;
 use Neos\FluidAdaptor\Core\ViewHelper\AbstractViewHelper;
-use Neos\Fusion\ViewHelpers\FusionContextTrait;
+use Neos\Fusion\FluidAdapter\ViewHelpers\FusionContextTrait;
 use Neos\Neos\Domain\Model\RenderingMode;
 
 /**
